requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.11.12
fastapi==0.121.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
python-multipart
